Conflicts:
	Models/Context.cs

+++ BLL/DocumentRepository.cs
@@ -1,22 +1,24 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Data.Entity;
 using TSS.Models;
 namespace TSS.BLL
 {
     public class DocumentRepository:Repository<Document,Guid>
     {
-        public static readonly DocumentRepository Repository = new DocumentRepository();
-        private DocumentRepository() { }
+        public DocumentRepository() { }
 
         public void Delete(Guid id, Action<string> onDeleted) {
             using (Context db=new Context()) {
                 Document obj = db.Documents.Find(id);
                 if (null!=obj) {
-                    db.Documents.Where(p => p.ParentId == obj.Id).ToList()
+                    if (obj.IsFolder==1) {
+                        db.Documents.Where(p => p.ParentId == obj.Id).ToList()
                         .ForEach(m => db.Documents.Remove(m));
-
-                    db.Documents.Remove(obj);                    
+                    }
+                    
+                    db.Documents.Remove(obj);
                     db.SaveChanges();
                     onDeleted(obj.Path);
                 }
@@ -89,9 +91,15 @@ namespace TSS.BLL
 
         public IList<Document> Search(string key, string specialty = null)
         {
+            if (string.IsNullOrWhiteSpace(key)) {
+                return null;
+            }
             using (Context db = new Context()) {
                 var query = db.Documents
-                    .Where(p => p.IsFolder == 0 && p.Name.ToLower().Contains(key.ToLower()) && (string.IsNullOrWhiteSpace(specialty) ? true : p.SpecialtyId == specialty));
+                    .Where(p => p.IsFolder == 0 && p.Name.ToLower().Contains(key.ToLower()));
+                if (null !=specialty) {
+                    query = query.Where(p => p.SpecialtyId.Equals(specialty));
+                }
                 return query.ToList();
             }
         }
